pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.57.1",
    "python-dotenv>=1.0.1",
    "streamlit>=1.40.0",
    "watchdog>=5.0.3",
]

[tool.uv.sources]
swarm = { git = "https://github.com/openai/swarm.git" }
